move closure lowering into expr::closure
